+++ internal/daemon/loop_api_payloads.go
@@ -97,7 +97,7 @@ func loopRunPayload(run looppkg.Run) (contract.LoopRunPayload, error) {
 		WorkspaceID:                  string(run.WorkspaceID),
 		LoopName:                     run.LoopName,
 		Status:                       contract.LoopRunStatus(run.Status),
-		Generation:                   run.Generation,
+		Generation:                   int64(run.Generation),
 		ReattemptStrategy:            contract.LoopReattemptStrategy(run.ReattemptStrategy),
 		CreatedAt:                    run.CreatedAt,
 		StartedAt:                    run.StartedAt,
